userKB/requirements.txt
SPARQLWrapper
JPype1
Flask